+++ src/CustomProperty.ts
@@ -110,8 +110,6 @@ export class CustomProperty {
     if (arg.choiceValues) customProperty.choiceValues = arg.choiceValues;
     if (arg.objectTypes) customProperty.objectTypes = arg.objectTypes;
     if (arg.valueType) customProperty.valueType = arg.valueType;
-    if (arg.modifiedByUserName)
-      customProperty.modifiedByUserName = arg.modifiedByUserName;
 
     customProperty.modifiedDate = modifiedDateTime();