+++ test/extension.spec.ts
@@ -1,0 +1,16 @@
+import chai from "chai";
+import { Config } from "./Config";
+
+const expect = chai.expect;
+const config = new Config();
+const repoApi = config.repoApi;
+
+describe("Extension operations", function () {
+  this.timeout(30000);
+
+  it("Export extension", async function () {
+    // let extension = await repoApi.extensionExport("basic-mashup-template");
+
+    expect(true).to.be.equal(true);
+  });
+});