+++ CHANGELOG.md
@@ -2,6 +2,28 @@
 
 All notable changes to this project will be documented in this file.
 
+## [0.0.3] - 2021-07-22
+
+### Added
+
+- `appUploadReplace` method
+- `genericRepoClient` exposed. Used for all Repository endpoints which dont have `/qrs` prefix (for example `/tempcontent` endpoints)
+- `genericWESClient` exposed. Used for `Web Extension Service API` endpoints. With port `9080`. For example `exportExtension` is one of these methods
+- `Phase 1` test are started
+
+### Changed
+
+- `appImport` method is renamed to `appUpload`
+- `appExport` correctly returns the app content as `Buffer`
+- `extensionExport` correctly returns the app content as `Buffer`
+- `tagRemoveFilter` throws an error when 0 items are returned
+- all `xxxRemove` methods are returning data in `IHttpReturnRemove` format (`{ id, status }`)
+
+### Removed
+
+- `exportExtension` method. This is not part of Repository APIs
+- `modifiedByUserName` parameter is no longer an option
+
 ## [0.0.2] - 2021-07-20
 
 ### Added